+++ packages/hightide-native/src/components/chat/ChatConversationRow.tsx
@@ -8,9 +8,7 @@ import {
 } from 'react'
 import {
   View,
-  type PressableProps,
-  type StyleProp,
-  type ViewStyle
+  type PressableProps
 } from 'react-native'
 import { HightideIconRegistry } from '../../icons/HightideIconRegistry'
 import { ThemedIcon } from '../visualization-and-display/ThemedIcon'
@@ -46,8 +44,7 @@ export type ChatConversationRowProps = Omit<PressableProps, 'children' | 'style'
   unreadCount?: number,
   isSelected?: boolean,
   sentIndicator?: ChatConversationSentIndicator,
-  style?: StyleProp<ViewStyle>,
-  rowStyle?: StyleOverwrite<PressableState, PressableContainerStyle>,
+  style?: StyleOverwrite<PressableState, PressableContainerStyle>,
   contentContainerStyle?: StyleOverwrite<ChatConversationRowState, ChatConversationRowContentContainerStyle>,
   headerRowStyle?: StyleOverwrite<ChatConversationRowState, ChatConversationRowHeaderRowStyle>,
   messageRowStyle?: StyleOverwrite<ChatConversationRowState, ChatConversationRowMessageRowStyle>,
@@ -83,7 +80,6 @@ export const ChatConversationRow = ({
   sentIndicator,
   disabled,
   style,
-  rowStyle,
   contentContainerStyle,
   headerRowStyle,
   messageRowStyle,
@@ -141,28 +137,28 @@ export const ChatConversationRow = ({
 
     return cloneElement(avatarElement, {
       size: avatarSize,
-      avatarStyle: (_, avatarState) => avatarTheme.container(
+      avatarStyle: (avatarState) => avatarTheme.container(
         {
           ...avatarState,
           size: avatarSize,
         },
         avatarStyle ?? avatarElement.props.avatarStyle
       ),
-      imageStyle: (_, avatarState) => avatarTheme.image(
+      imageStyle: (avatarState) => avatarTheme.image(
         {
           ...avatarState,
           size: avatarSize,
         },
         avatarElement.props.imageStyle
       ),
-      textStyle: (_, avatarState) => avatarTheme.text(
+      textStyle: (avatarState) => avatarTheme.text(
         {
           ...avatarState,
           size: avatarSize,
         },
         avatarElement.props.textStyle
       ),
-      iconStyle: (_, avatarState) => avatarTheme.icon(
+      iconStyle: (avatarState) => avatarTheme.icon(
         {
           ...avatarState,
           size: avatarSize,
@@ -176,23 +172,22 @@ export const ChatConversationRow = ({
     <ThemedPressable
       {...props}
       disabled={disabled}
-      style={style}
-      containerStyle={(_, pressableState) => {
+      style={(pressableState) => {
         const state = resolveState(pressableState as PressableInteraction)
         return theme.components.chat.conversationRow.pressable(state).container(
           pressableState,
-          rowStyle
+          style
         )
       }}
-      stateLayerStyle={(_, pressableState) => {
+      stateLayerStyle={(pressableState) => {
         const state = resolveState(pressableState as PressableInteraction)
         return theme.components.chat.conversationRow.pressable(state).stateLayer(pressableState)
       }}
-      textStyle={(_, pressableState) => {
+      textStyle={(pressableState) => {
         const state = resolveState(pressableState as PressableInteraction)
         return theme.components.chat.conversationRow.pressable(state).text(pressableState)
       }}
-      iconStyle={(_, pressableState) => {
+      iconStyle={(pressableState) => {
         const state = resolveState(pressableState as PressableInteraction)
         return theme.components.chat.conversationRow.pressable(state).icon(pressableState)
       }}
@@ -228,21 +223,23 @@ export const ChatConversationRow = ({
                 )}
               </View>
               <View style={resolvedMessageRow}>
-                {sentIndicator && (
-                  <ThemedIcon
-                    icon={sentIndicatorIcon}
-                    size={resolvedSentIndicator.size}
-                    strokeWidth={resolvedSentIndicator.strokeWidth}
-                    color={resolvedSentIndicator.color}
-                  />
-                )}
-                {preview != null && (
-                  typeof preview === 'string' || typeof preview === 'number' ? (
-                    <ThemedText style={resolvedPreview} numberOfLines={1}>{preview}</ThemedText>
-                  ) : (
-                    preview
-                  )
-                )}
+                <View style={{ flexDirection: 'row', alignItems: 'center', flex: 1, gap: theme.spacing.xs  /* TODO replace this with a theme style */ }}>
+                  {sentIndicator && (
+                    <ThemedIcon
+                      icon={sentIndicatorIcon}
+                      size={resolvedSentIndicator.size}
+                      strokeWidth={resolvedSentIndicator.strokeWidth}
+                      color={resolvedSentIndicator.color}
+                    />
+                  )}
+                  {preview != null && (
+                    typeof preview === 'string' || typeof preview === 'number' ? (
+                      <ThemedText style={resolvedPreview} numberOfLines={1}>{preview}</ThemedText>
+                    ) : (
+                      preview
+                    )
+                  )}
+                </View>
                 {isUnread && (
                   <View style={unreadBadge}>
                     <ThemedText style={unreadBadgeText}>{unreadCount}</ThemedText>