+++ packages/hightide-native/src/global-contexts/content-theme/ContentThemeProvider.tsx
@@ -16,14 +16,16 @@ export type ContentThemeRootProviderProps = PropsWithChildren & ContentThemeCont
 export const ContentThemeRootProvider = ({
   children,
   foreground,
+  background,
   textStyle,
   iconStyle,
 }: ContentThemeRootProviderProps) => {
   const value = useMemo((): ContentThemeContextValue => ({
     foreground,
+    background,
     textStyle,
     iconStyle,
-  }), [foreground, textStyle, iconStyle])
+  }), [foreground, background, textStyle, iconStyle])
 
   return (
     <ContentThemeContext.Provider value={value}>
@@ -32,7 +34,7 @@ export const ContentThemeRootProvider = ({
   )
 }
 
-type ForegroundOverride =
+type ColorOverride =
   | ColorValue
   | ((prev: ColorValue) => ColorValue)
 
@@ -45,38 +47,53 @@ type IconStyleOverride =
   | ((prev: IconStyle, foreground: ColorValue) => IconStyle)
 
 export type ContentThemeOverrideProviderProps = PropsWithChildren & {
-  foreground?: ForegroundOverride,
+  foreground?: ColorOverride,
+  background?: ColorOverride,
   textStyle?: TextStyleOverride,
   iconStyle?: IconStyleOverride,
+  isKeepingTextColor?: boolean,
+  isKeepingIconColor?: boolean,
 }
 
 export const ContentThemeOverrideProvider = ({
   children,
   foreground: foregroundOverride,
+  background: backroundOverride,
   textStyle: textStyleOverride,
   iconStyle: iconStyleOverride,
+  isKeepingIconColor = false,
+  isKeepingTextColor = false,
 }: ContentThemeOverrideProviderProps) => {
   const parent = useContentTheme()
 
   const value = useMemo((): ContentThemeContextValue => {
-    const foreground = typeof foregroundOverride === 'function'
+    const newForeground = typeof foregroundOverride === 'function'
       ? foregroundOverride(parent.foreground)
-      : foregroundOverride ?? parent.foreground
+      : foregroundOverride
+
+    const newBackground = typeof backroundOverride === 'function'
+      ? backroundOverride(parent.background)
+      : backroundOverride
 
     const textStyle = typeof textStyleOverride === 'function'
-      ? textStyleOverride(parent.textStyle, foreground)
+      ? textStyleOverride(parent.textStyle, parent.foreground)
       : textStyleOverride ?? parent.textStyle
 
+    if(!isKeepingTextColor && newForeground !== undefined) textStyle['color'] = newForeground
+
     const iconStyle = typeof iconStyleOverride === 'function'
-      ? iconStyleOverride(parent.iconStyle, foreground)
+      ? iconStyleOverride(parent.iconStyle, parent.foreground)
       : iconStyleOverride ?? parent.iconStyle
 
+    if(!isKeepingIconColor && newForeground !== undefined) iconStyle['color'] = newForeground
+
     return {
-      foreground,
+      foreground: newForeground ?? parent.foreground,
+      background: newBackground ?? parent.background,
       textStyle,
       iconStyle,
     }
-  }, [parent, foregroundOverride, textStyleOverride, iconStyleOverride])
+  }, [parent, foregroundOverride, textStyleOverride, iconStyleOverride, backroundOverride, isKeepingTextColor, isKeepingIconColor])
 
   return (
     <ContentThemeContext.Provider value={value}>